Makes the read/write_bare functions available through a submodule.
Documentation has been updated to reflect this (and other) changes,
and the code has been reorganized into separate files.
